feat(vscode): Add tasks.json to speed up dev build
As a good part of the team is using VS Code, I think it is useful to
execute just one command to get the project ready to work.

The default shortcut in VS Code is `Ctrl` + `Shift` + `B`. In my case, I
assigned it to the `F6` key. It is really quite useful.